files restructured and removed unnecessary comments
