+++ src/App.js
@@ -1,7 +1,7 @@
-import DiaryHome from './page/diary-page/diary-page';
+import DiaryHome from './pages/diary-page/diary-page';
 import './App.css';
 import { Provider } from 'react-redux';
-import { store } from './state/store';
+import { store } from './utils/state/store';
 
 function App() {
